Revert "preload sections content for a better menu experience"
