+++ core/blocks/sections/index.jsx
@@ -9,11 +9,42 @@ import Groups from "./Groups"
 
 const map = (state) => ({ sections: state.sections })
 
+const bindSections = (props) => {
+  const { dispatch } = props
+
+  let handle = {}
+  Tracker.autorun((computation) => {
+    // return computation for dismount
+    handle = computation
+
+    // subscribe to sections
+    Meteor.subscribe("sections")
+
+    // pull all the sections data
+    let sections = Sections.find({}, { sort: { order: 1 }}).fetch()
+
+    let _sections = {}
+    for (let _item of sections) { _sections[_item._id] = _item }
+
+    // persist in the store
+    dispatch(sectionActions.set(_sections))
+
+  })
+
+  return { handle }
+
+}
 
 @connect(map)
 export default class SectionsContainer extends Component {
 
+  componentWillMount(){
+    let { handle } = bindSections(this.props)
+    this.handle = handle
+  }
+
   componentWillUnmount(){
+    this.handle.stop()
     this.props.dispatch(modal.update({keepNav: false}))
   }
